+++ spec/support/servers/modules/recon/localstart_asp.rb
@@ -1,0 +1,5 @@
+require 'sinatra'
+
+get '/localstart.asp' do
+    status 401
+end